# yd-play-v1.0-stable/docs/openapi.yaml
openapi: 3.1.0
info:
  title: YD Play API
  version: 1.0.0
  description: 'Auth, virtual-wallet, referrals, anti-abuse, and the Pulse Grid authoritative single-player game milestone for YD Play. Coins are virtual, non-transferable, have no real-world cash value, and cannot be cashed out.


    v0.5 adds read-only multiplayer room status endpoints. Realtime room mutations and gameplay use the authenticated WebSocket protocol documented in docs/REALTIME.md.


    v0.9 adds missions, achievements, validated Pulse Grid leaderboards, authenticated analytics events, in-app notifications, provider-neutral push-device registration, campaign dispatch, and mobile safety/report history.

    v1.0.0 adds release hardening, production configuration guards, and separate liveness/readiness health endpoints.'
servers:
- url: http://localhost:3000/v1
paths:
...
  /auth/login:
    post:
      summary: Login and create a refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or inactive account
...
  /auth/logout:
    post:
      security:
      - bearerAuth: []
      summary: Revoke current session
      responses:
        '201':
          description: Revoked
...
  /admin/referrals/config:
    get:
      summary: Read referral reward, cap and fraud configuration
      parameters:
      - $ref: '#/components/parameters/AdminApiKey'
      responses:
        '200':
          description: Referral configuration
        '401':
          description: Invalid admin key
    patch:
      summary: Update referral reward, cap and fraud configuration
      parameters:
      - $ref: '#/components/parameters/AdminApiKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferralConfigPatch'
      responses:
        '200':
          description: Updated configuration
  /admin/referrals/fraud-flags:
    get:
      summary: List referral abuse flags and evidence
      parameters:
      - $ref: '#/components/parameters/AdminApiKey'
      - name: status
        in: query
        schema:
          type: string
          enum:
          - open
          - reviewing
          - confirmed
          - dismissed
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: Fraud flag page
  /admin/referrals/fraud-flags/{flagId}:
    patch:
      summary: Review a fraud flag
      parameters:
      - $ref: '#/components/parameters/AdminApiKey'
      - name: flagId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewFlagRequest'
      responses:
        '200':
          description: Updated fraud flag
  /admin/referrals/{referralId}/decision:
    patch:
      summary: Approve or reject a referral after manual evidence review
      parameters:
      - $ref: '#/components/parameters/AdminApiKey'
      - name: referralId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferralDecisionRequest'
      responses:
        '200':
          description: Referral decision result
  /admin/referrals/overview:
    get:
      summary: Referral totals, review counts, rejected counts and granted coins
      parameters:
      - $ref: '#/components/parameters/AdminApiKey'
      responses:
        '200':
          description: Referral overview
  /games:
    get:
      security:
      - bearerAuth: []
      summary: List active games
      responses:
        '200':
          description: Active games
  /games/pulse-grid/sessions:
    post:
      security:
      - bearerAuth: []
      summary: Start an authoritative Pulse Grid session
      responses:
        '201':
          description: Game session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSessionStart'
  /games/sessions/{sessionId}:
    get:
      security:
      - bearerAuth: []
      summary: Get authoritative game-session state
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Game session state
        '404':
          description: Session not found
  /games/sessions/{sessionId}/cues:
    get:
      security:
      - bearerAuth: []
      summary: Release only near-future Pulse Grid target cues
      parameters:
      - $ref: '#/components/parameters/SessionId'
      - name: after
        in: query
        required: false
        schema:
          type: integer
          minimum: -1
          default: -1
      responses:
        '200':
          description: Short look-ahead cue window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameCueBatch'
  /games/sessions/{sessionId}/input:
    post:
      security:
      - bearerAuth: []
      summary: Submit one ordered player input for server validation
      parameters:
      - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameInputRequest'
      responses:
        '201':
          description: Input accepted and authoritative score returned
        '409':
          description: Stale sequence, future input, late input, or inactive session
  /games/sessions/{sessionId}/complete:
    post:
      security:
      - bearerAuth: []
      summary: Replay-validate and finalize a game, then grant an idempotent virtual-coin reward
      parameters:
      - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameCompleteRequest'
      responses:
        '201':
          description: Authoritative result and reward status
        '409':
          description: Game is still in progress or cannot be completed
...
  /admin/auth/login:
    post:
      tags:
      - Admin
      security: []
      summary: Login with a named admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 10
      responses:
        '201':
          description: Admin access token
        '401':
          description: Invalid credentials
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AdminApiKey:
      name: X-Admin-Api-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 24
      description: Temporary bootstrap admin credential for v0.3; replace with named admin MFA/RBAC.
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ReferralConfigPatch:
      type: object
      additionalProperties: false
      properties:
        inviterReward:
          type: integer
          minimum: 0
        inviteeReward:
          type: integer
          minimum: 0
        dailyInviterRewardCap:
          type: integer
          minimum: 0
        lifetimeInviterRewardCap:
          type: integer
          minimum: 0
        minAccountAgeMinutes:
          type: integer
          minimum: 0
        maxInvitesPerHour:
          type: integer
          minimum: 1
        sameDeviceRisk:
          type: integer
          minimum: 0
          maximum: 100
        sameIpRisk:
          type: integer
          minimum: 0
          maximum: 100
        ipVelocityRisk:
          type: integer
          minimum: 0
          maximum: 100
        deviceMultiAccountRisk:
          type: integer
          minimum: 0
          maximum: 100
        inviterVelocityRisk:
          type: integer
          minimum: 0
          maximum: 100
        reviewThreshold:
          type: integer
          minimum: 0
          maximum: 100
        blockThreshold:
          type: integer
          minimum: 0
          maximum: 100
    ReviewFlagRequest:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - reviewing
          - confirmed
          - dismissed
        reason:
          type: string
          minLength: 3
          maxLength: 300
    ReferralDecisionRequest:
      type: object
      required:
      - decision
      - reason
      properties:
        decision:
          type: string
          enum:
          - approve
          - reject
        reason:
          type: string
          minLength: 3
          maxLength: 300
...
    LoginRequest:
      type: object
      required:
      - identifier
      - password
      - platform
      - deviceProof
      properties:
        identifier:
          type: string
        password:
          type: string
          minLength: 10
        platform:
          type: string
          enum:
          - android
          - ios
          - web
        deviceProof:
          type: string
        appVersion:
          type: string
    AuthResponse:
      type: object
      required:
      - accessToken
      - refreshToken
      - user
      - wallet
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          additionalProperties: true
        wallet:
          $ref: '#/components/schemas/Wallet'
    Wallet:
      type: object
      required:
      - currencyCode
      - balance
      - cashValue
      - transferable
      - cashOut
      properties:
        currencyCode:
          type: string
          const: COIN
        balance:
          type: string
          description: Integer string to preserve BIGINT precision across clients
        cashValue:
          type: integer
          const: 0
        transferable:
          type: boolean
          const: false
        cashOut:
          type: boolean
          const: false
    PulseTarget:
      type: object
      required:
      - index
      - timeMs
      - lane
      properties:
        index:
          type: integer
          minimum: 0
        timeMs:
          type: integer
          minimum: 0
        lane:
          type: integer
          minimum: 0
          maximum: 3
    GameSessionStart:
      type: object
      required:
      - id
      - game
      - status
      - startedAt
      - serverNow
      - durationMs
      - hitWindowMs
      - lanes
      - cueLookaheadMs
      - scheduleHash
      properties:
        id:
          type: string
          format: uuid
        game:
          type: object
          additionalProperties: true
        status:
          type: string
          const: active
        startedAt:
          type: string
          format: date-time
        serverNow:
          type: string
          format: date-time
        durationMs:
          type: integer
          const: 45000
        hitWindowMs:
          type: integer
        lanes:
          type: integer
          const: 4
        cueLookaheadMs:
          type: integer
          const: 1200
        scheduleHash:
          type: string
        disclaimer:
          type: string
    GameCueBatch:
      type: object
      required:
      - items
      - serverNow
      - serverElapsedMs
      - releaseThroughMs
      - scheduleHash
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PulseTarget'
        serverNow:
          type: string
          format: date-time
        serverElapsedMs:
          type: integer
          minimum: 0
        releaseThroughMs:
          type: integer
          minimum: 0
        scheduleHash:
          type: string
    GameInputRequest:
      type: object
      additionalProperties: false
      required:
      - seq
      - elapsedMs
      - lane
      properties:
        seq:
          type: integer
          minimum: 1
        elapsedMs:
          type: integer
          minimum: 0
          maximum: 120000
        lane:
          type: integer
          minimum: 0
          maximum: 3
    GameCompleteRequest:
      type: object
      additionalProperties: false
      properties:
        claimedScore:
          type: integer
          minimum: 0
          description: Optional client display score. Server score remains authoritative even when this differs.
...